% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsi.R
\name{fsi_countries}
\alias{fsi_countries}
\title{fsi_countries}
\usage{
fsi_countries()
}
\value{
list of countries
}
\description{
This function provide a list of countries in FSI database
}
\examples{
fsi_countries()
}
