% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsi.R
\name{read_fsi}
\alias{read_fsi}
\title{read_fsi}
\usage{
read_fsi(
  startyear = NULL,
  endyear = NULL,
  country = NULL,
  indicator = NULL,
  rank = TRUE,
  total = TRUE
)
}
\arguments{
\item{startyear}{start year of extraction, the default is NULL}

\item{endyear}{end year of extraction, the default is NULL}

\item{country}{a specific country or list of countries, see fsi_countries}

\item{indicator}{a list of FSI indicator, see fsi_series}

\item{rank}{extract the rank, default is TRUE}

\item{total}{extract the Total, default is TRUE}
}
\value{
FSI data.frame
}
\description{
read_fsi
}
\examples{
read_fsi(startyear=2020, endyear=2021)
}
